+++ src/sas/readstat_sas7bcat_read.c
@@ -90,7 +90,6 @@ static readstat_error_t sas7bcat_parse_value_labels(const char *value_start, siz
             goto cleanup;
         }
         readstat_value_t value = { .type = is_string ? READSTAT_TYPE_STRING : READSTAT_TYPE_DOUBLE };
-        size_t label_len = sas_read2(&lbp2[8], ctx->bswap);
         char string_val[4*16+1];
         if (is_string) {
             size_t value_entry_len = 6 + lbp1[2];
@@ -104,19 +103,18 @@ static readstat_error_t sas7bcat_parse_value_labels(const char *value_start, siz
             uint64_t val = sas_read8(&lbp1[22], bswap_doubles);
             double dval = NAN;
             if ((val | 0xFF0000000000) == 0xFFFFFFFFFFFF) {
-                value.tag = (val >> 40);
-                if (sas_validate_tag(value.tag) == READSTAT_OK) {
-                    value.is_tagged_missing = 1;
-                } else {
-                    value.tag = 0;
-                    value.is_system_missing = 1;
-                }
+                sas_assign_tag(&value, (val >> 40));
             } else {
                 memcpy(&dval, &val, 8);
                 dval *= -1.0;
             }
 
             value.v.double_value = dval;
+        }
+        size_t label_len = sas_read2(&lbp2[8], ctx->bswap);
+        if (&lbp2[10] + label_len - value_start > value_labels_len) {
+            retval = READSTAT_ERROR_PARSE;
+            goto cleanup;
         }
         if (ctx->value_label_handler) {
             char label[4*label_len+1];
@@ -150,7 +148,7 @@ static readstat_error_t sas7bcat_parse_block(const char *data, size_t data_size,
     int payload_offset = 106;
     char name[4*32+1];
 
-    if (data_size < 50)
+    if (data_size < payload_offset)
         goto cleanup;
 
     pad = (data[2] & 0x08) ? 4 : 0; // might be 0x10, not sure